move alerts to collection detail
