+++ src/components/collection-detail/collection-info.tsx
@@ -34,6 +34,7 @@ interface IProps extends CollectionDetailType {
     version?: string;
   };
   updateParams: (params) => void;
+  addAlert?: (variant, title, description?) => void;
 }
 
 interface IState {
@@ -65,122 +66,114 @@ export class CollectionInfo extends React.Component<IProps, IState> {
     }
 
     return (
-      <>
-        <AlertList
-          alerts={this.state.alerts}
-          closeAlert={(i) => this.closeAlert(i)}
-        />
-        <div className='pf-c-content info-panel'>
-          <h1>{t`Install`}</h1>
-          <Grid hasGutter={true}>
-            <GridItem>{latest_version.metadata.description}</GridItem>
-            <GridItem>
-              {latest_version.metadata.tags.map((tag, i) => (
-                <Tag key={i}>{tag}</Tag>
-              ))}
-            </GridItem>
+      <div className='pf-c-content info-panel'>
+        <h1>{t`Install`}</h1>
+        <Grid hasGutter={true}>
+          <GridItem>{latest_version.metadata.description}</GridItem>
+          <GridItem>
+            {latest_version.metadata.tags.map((tag, i) => (
+              <Tag key={i}>{tag}</Tag>
+            ))}
+          </GridItem>
 
-            <GridItem>
-              <Split hasGutter={true}>
-                <SplitItem className='install-title'>{t`License`}</SplitItem>
-                <SplitItem isFilled>
-                  {latest_version.metadata.license}
-                </SplitItem>
-              </Split>
-            </GridItem>
-            <GridItem>
-              <Split hasGutter={true}>
-                <SplitItem className='install-title'>{t`Installation`}</SplitItem>
-                <SplitItem isFilled>
-                  <ClipboardCopy isReadOnly>{installCommand}</ClipboardCopy>
-                  <div>
-                    <Trans>
-                      <b>Note:</b> Installing collections with ansible-galaxy is
-                      only supported in ansible 2.9+
-                    </Trans>
-                  </div>
-                  {!this.context.settings
-                    .GALAXY_ENABLE_UNAUTHENTICATED_COLLECTION_DOWNLOAD &&
-                  this.context.user.is_anonymous ? (
-                    <Alert
-                      className={'collection-download-alert'}
-                      isInline
-                      variant='warning'
-                      title={
-                        <React.Fragment>
-                          {t`You have to be logged in to be able to download the tarball.`}{' '}
-                          <LoginLink />
-                        </React.Fragment>
-                      }
-                    />
-                  ) : (
-                    <div>
-                      <a
-                        ref={this.downloadLinkRef}
-                        style={{ display: 'none' }}
-                      ></a>
-                      <Button
-                        className='download-button'
-                        variant='link'
-                        icon={<DownloadIcon />}
-                        onClick={() =>
-                          this.download(
-                            this.context.selectedRepo,
-                            namespace,
-                            name,
-                            latest_version,
-                          )
-                        }
-                      >
-                        {t`Download tarball`}
-                      </Button>
-                    </div>
-                  )}
-                </SplitItem>
-              </Split>
-            </GridItem>
-            {latest_version.requires_ansible && (
-              <GridItem>
-                <Split hasGutter={true}>
-                  <SplitItem className='install-title'>
-                    {t`Requires Ansible`}
-                  </SplitItem>
-                  <SplitItem isFilled>
-                    {latest_version.requires_ansible}
-                  </SplitItem>
-                </Split>
-              </GridItem>
-            )}
-
-            {latest_version.docs_blob.collection_readme ? (
-              <GridItem>
-                <div className='readme-container'>
-                  <div
-                    className='pf-c-content'
-                    dangerouslySetInnerHTML={{
-                      __html: latest_version.docs_blob.collection_readme.html,
-                    }}
-                  />
-                  <div className='fade-out'></div>
+          <GridItem>
+            <Split hasGutter={true}>
+              <SplitItem className='install-title'>{t`License`}</SplitItem>
+              <SplitItem isFilled>{latest_version.metadata.license}</SplitItem>
+            </Split>
+          </GridItem>
+          <GridItem>
+            <Split hasGutter={true}>
+              <SplitItem className='install-title'>{t`Installation`}</SplitItem>
+              <SplitItem isFilled>
+                <ClipboardCopy isReadOnly>{installCommand}</ClipboardCopy>
+                <div>
+                  <Trans>
+                    <b>Note:</b> Installing collections with ansible-galaxy is
+                    only supported in ansible 2.9+
+                  </Trans>
                 </div>
-                <Link
-                  to={formatPath(
-                    Paths.collectionDocsIndexByRepo,
-                    {
-                      collection: name,
-                      namespace: namespace.name,
-                      repo: this.context.selectedRepo,
-                    },
-                    params,
-                  )}
-                >
-                  {t`Go to documentation`}
-                </Link>
-              </GridItem>
-            ) : null}
-          </Grid>
-        </div>
-      </>
+                {!this.context.settings
+                  .GALAXY_ENABLE_UNAUTHENTICATED_COLLECTION_DOWNLOAD &&
+                this.context.user.is_anonymous ? (
+                  <Alert
+                    className={'collection-download-alert'}
+                    isInline
+                    variant='warning'
+                    title={
+                      <React.Fragment>
+                        {t`You have to be logged in to be able to download the tarball.`}{' '}
+                        <LoginLink />
+                      </React.Fragment>
+                    }
+                  />
+                ) : (
+                  <div>
+                    <a
+                      ref={this.downloadLinkRef}
+                      style={{ display: 'none' }}
+                    ></a>
+                    <Button
+                      className='download-button'
+                      variant='link'
+                      icon={<DownloadIcon />}
+                      onClick={() =>
+                        this.download(
+                          this.context.selectedRepo,
+                          namespace,
+                          name,
+                          latest_version,
+                        )
+                      }
+                    >
+                      {t`Download tarball`}
+                    </Button>
+                  </div>
+                )}
+              </SplitItem>
+            </Split>
+          </GridItem>
+          {latest_version.requires_ansible && (
+            <GridItem>
+              <Split hasGutter={true}>
+                <SplitItem className='install-title'>
+                  {t`Requires Ansible`}
+                </SplitItem>
+                <SplitItem isFilled>
+                  {latest_version.requires_ansible}
+                </SplitItem>
+              </Split>
+            </GridItem>
+          )}
+
+          {latest_version.docs_blob.collection_readme ? (
+            <GridItem>
+              <div className='readme-container'>
+                <div
+                  className='pf-c-content'
+                  dangerouslySetInnerHTML={{
+                    __html: latest_version.docs_blob.collection_readme.html,
+                  }}
+                />
+                <div className='fade-out'></div>
+              </div>
+              <Link
+                to={formatPath(
+                  Paths.collectionDocsIndexByRepo,
+                  {
+                    collection: name,
+                    namespace: namespace.name,
+                    repo: this.context.selectedRepo,
+                  },
+                  params,
+                )}
+              >
+                {t`Go to documentation`}
+              </Link>
+            </GridItem>
+          ) : null}
+        </Grid>
+      </div>
     );
   }
 
@@ -201,20 +194,11 @@ export class CollectionInfo extends React.Component<IProps, IState> {
         this.downloadLinkRef.current.click();
       })
       .catch((e) =>
-        this.setState({
-          alerts: [
-            ...this.state.alerts,
-            {
-              variant: 'danger',
-              title: t`Error downloading collection.`,
-              description: e?.message,
-            },
-          ],
-        }),
+        this.props.addAlert(
+          'danger',
+          t`Error downloading collection.`,
+          e?.message,
+        ),
       );
-  }
-
-  private get closeAlert() {
-    return closeAlertMixin('alerts');
   }
 }
